@@ -10,6 +10,7 @@
     "Productivity"
   ],
   "license": "MIT",
+  "type": "module",
   "commands": [
     {
       "name": "r2-uploader",
@@ -114,7 +115,7 @@
     "@raycast/eslint-config": "^2.0.4",
     "@types/node": "22.13.10",
     "@types/react": "19.0.10",
-    "eslint": "^9.22.0",
+    "eslint": "^9.34.0",
     "prettier": "^3.5.3",
     "typescript": "^5.8.2"
   },
